requirements.txt
# Django dependencies
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0



# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Swagger/API documentation
drf-yasg==1.21.10
swagger-spec-validator==3.0.4

# Authentication & Security
PyJWT==2.9.0
passlib==1.7.4
python-multipart==0.0.12

# General utilities
python-dotenv==1.1.0
Pillow==11.2.1
pytz==2025.2

# JSON/Schema handling
attrs==25.3.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1

# Other dependencies
importlib_resources==6.5.2
inflection==0.5.1
Markdown==3.8
packaging==25.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0